Fixed The Steering Wheel WOOOO!!
Steering wheel of the ship now supports steering with 2 hands (unstable), and with 1 hand.
Auto detaches hands when they are not close enough to the steering wheel.
Steering wheel acts like it has velocity based movement (keeps spinning after spun left hard).
Steering wheel spins left and right randomly because of the waves in the water (only when not held).