Updated Steering Wheel
When holding the wheel with 2 hands, both hands contribute half to the rotation of the wheel > this works great ;P
Wheel now Rotates back to center on top of all rotation logic (when player doesnt hold the wheel).
Wheel now also has a random chance every 30 seconds to get a HUGE boost of rotational speed to 1 direction so the wheel flies to 1 side.